Improve repositories, parsing, and ViewModel data handling

+++ app/src/main/java/com/mkulimamarket/app/domain/model/PriceTrend.kt
@@ -2,5 +2,5 @@ package com.mkulimamarket.app.domain.model
 
 data class PriceTrend(
     val week: String,
-    val price: Int
+    val price: Double
 )
